requirements.txt
# Streamlit Cloudの環境と互換性の高い代替パッケージ
google-cloud-aiplatformgoogle-generativeai
streamlit
google-generativeai  <-- この名前が正解です

google-genai
google-generativeai==0.1.0
# Streamlit Cloudの環境と互換性の高い代替パッケージ
google-cloud-aiplatform

# Streamlit Cloudの環境と互換性の高い代替パッケージ
google-cloud-aiplatform
